+++ strucmotif-search-core/src/main/java/org/rcsb/strucmotif/io/AssemblyInformationProviderImpl.java
@@ -1,6 +1,5 @@
 package org.rcsb.strucmotif.io;
 
-import org.rcsb.strucmotif.config.StrucmotifConfig;
 import org.rcsb.strucmotif.domain.Pair;
 import org.rcsb.strucmotif.domain.structure.StructureInformation;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,17 +22,14 @@ public class AssemblyInformationProviderImpl implements AssemblyInformationProvi
     private static final Map<String, Set<String>> UNDEFINED_ASSEMBLIES = Collections.emptyMap();
     // maps from struct_oper_id to all assemblies this transformation is part of
     private final Map<String, Map<String, Set<String>>> reverseAssemblyInformation;
-    private final StrucmotifConfig strucmotifConfig;
 
     /**
      * Construct an assembly info provider.
      * @param stateRepository the state provider
-     * @param strucmotifConfig the config
      */
     @Autowired
-    public AssemblyInformationProviderImpl(StateRepository stateRepository, StrucmotifConfig strucmotifConfig) {
+    public AssemblyInformationProviderImpl(StateRepository stateRepository) {
         this.reverseAssemblyInformation = loadAssemblyInformation(stateRepository);
-        this.strucmotifConfig = strucmotifConfig;
     }
 
     private Map<String, Map<String, Set<String>>> loadAssemblyInformation(StateRepository stateRepository) {
@@ -64,10 +60,5 @@ public class AssemblyInformationProviderImpl implements AssemblyInformationProvi
     @Override
     public Map<String, Set<String>> selectAssemblyMap(String structureIdentifier) {
         return reverseAssemblyInformation.getOrDefault(structureIdentifier, UNDEFINED_ASSEMBLIES);
-    }
-
-    @Override
-    public String getUndefinedAssemblyIdentifier() {
-        return strucmotifConfig.getUndefinedAssemblyIdentifier();
     }
 }