+++ strucmotif-search-core/src/test/java/org/rcsb/strucmotif/domain/structure/ConnectednessTest.java
@@ -23,7 +23,6 @@ public class ConnectednessTest {
     public void init() {
         strucmotifConfig = new StrucmotifConfig();
         strucmotifConfig.setDistanceCutoff(TEST_DISTANCE_CUTOFF);
-        strucmotifConfig.setUndefinedAssemblies(false);
         structureReader = new StructureReaderImpl(new ResidueTypeResolverImpl(strucmotifConfig));
     }