+++ strucmotif-search-update/src/test/java/org/rcsb/strucmotif/update/TestCases.java
@@ -14,8 +14,8 @@ enum TestCases {
     PDB_4TUT("4TUT.bcif.gz"),
     PDB_5XES("5xes.bcif"),
     PDB_6FCE("6fce.cif"),
-    AF_A0A0R0FWM3("AF-A0A0R0FWM3-F1-model_v1.cif"), // helix, fragmented
-    AF_Q8SY76("af-q8sy76-f1-model_v1.cif"); // helix
+    AF_A0A0R0FWM3("af_afa0a0r0fwm3f1.bcif"), // helix, fragmented
+    AF_Q8SY76("af_afq8sy76f1.bcif"); // helix
 
     private final String filename;
     private final String key;